fix(core): test updates
Part of ENG-1067
